finished refactoring pets
